Fixed migration

+++ src/main/java/com/hotels/properties/models/Property.java
@@ -17,7 +17,7 @@ public class Property {
     String description;
     PropertyType propertyType;
     String listingCurrency;
-    Double nightlyRate;
+    Integer nightlyRate;
     Integer bathrooms;
     Boolean published;
     String address;
